+++ pkg/innerring/config/config.go
@@ -133,8 +133,9 @@ type Experimental struct {
 
 // Mainnet configures mainnet chain settings.
 type Mainnet struct {
-	Enabled               bool `mapstructure:"enabled"`
-	DisableGovernanceSync bool `mapstructure:"disable_governance_sync"`
+	Enabled               bool  `mapstructure:"enabled"`
+	DisableGovernanceSync bool  `mapstructure:"disable_governance_sync"`
+	ExtraFee              int64 `mapstructure:"extra_fee"`
 	BasicChain            `mapstructure:",squash"`
 }