+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog for NeoFS Node
 - Move `fschain_autodeploy` into `fschain.disable_autodeploy` in IR config (#3619)
 - Move `without_mainnet` into `mainnet.enabled` in IR config (#3619)
 - Move `governance.disable` into `mainnet.disable_governance_sync` in IR config (#3619)
+- Move `fee.main_chain` into `mainnet.extra_fee` in IR config (#3619)
 
 ### Removed
 
@@ -23,6 +24,7 @@ by default autodeploy is enabled now;
 - `mainnet.enabled` instead of deprecated `without_mainnet` with reverted value,
 by default mainnet is disabled now;
 - `mainnet.disable_governance_sync` instead of deprecated `governance.disable`;
+- `mainnet.extra_fee` instead of deprecated `fee.main_chain`;
 Old options are still supported but will be removed in future releases.
 
 ## [0.49.0] - 2025-10-06 - Dochodo